requirements.txt
anyio==3.7.1
asgiref==3.7.2
bidict==0.22.1
bleach==6.0.0
blinker==1.6.2
boto3==1.28.64
botocore==1.31.64
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.5
click-log==0.4.0
colorama==0.4.6
deprecation==2.1.0
distlib==0.3.6
dnspython==2.4.2
docutils==0.20.1
dotty-dict==1.3.1
feedparser==6.0.10
filelock==3.12.2
Flask==2.3.2
Flask-APScheduler==1.13.0
Flask-Cors==4.0.0
Flask-Script==2.0.6
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
gitdb==4.0.10
GitPython==3.1.32
gotrue==1.0.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
image==1.5.33
importlib-metadata==6.8.0
imutils==0.5.4
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jmespath==1.0.1
keyring==24.2.0
lxml==4.9.3
MarkupSafe==2.1.3
more-itertools==10.0.0
packaging==23.1
Pillow==10.0.0
pipenv==2023.7.23
pkginfo==1.9.6
platformdirs==3.8.1
postgrest==0.10.6
pydantic==1.10.12
Pygments==2.15.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.7.0
python-gitlab==3.15.0
python-semantic-release==7.33.2
python-socketio==5.9.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
readme-renderer==40.0
realtime==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
s3transfer==0.7.0
semver==2.13.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.4
virtualenv==20.23.1
virtualenv-clone==0.5.7
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.6
zipp==3.16.2